Remove mobile sidebar tutorial support
At one point we had tutorial steps that required the user to open the
sidebar and navigate to a different section. That was replaced with
steps that auto-navigate to the right section when needed.

This commit removes a ton of hacky code that supported opening the
sidebar from a tutorial step and managing an overlaid tutorial resume
button.