Tooltip nudge to select account for sidebar links
When the user isn't in an account context (e.g. in User Settings or the
Create Account screen), the sidebar links to account-specific screens
are disabled. It's not obvious why they are disabled, though.

This commit adds a tooltip that briefly displays when clicking on a
disabled link, prompting the user to select an account first.

Closes: Archodex/archodex#3