Fix resuming tutorials after browser reload
This fixes an issue where reloading the playground on later tutorial
steps could put the tutorial into a broken state where no tutorial
content is shown and the Resume Tutorial button would also not be shown.

On a browser reload where the tutorial was previously disabled we load
again with the tutorial disabled. But we don't save the reference to the
queryDataDispatch function properly in that edge case. This change
ensures the function ref is available in all cases.